REFACTOR: extracted logic to view
